refactor(MainActivity): Activity는 View로만 사용하도록 리스너 위치 변경
